Update Web Admin 26/5/2021

+++ Medical.Entities/DomainEntity/MedicalAppDomain.cs
@@ -18,7 +18,7 @@ namespace Medical.Entities.DomainEntity
         /// STT
         /// </summary>
         [NotMapped]
-        public int RowNumber { get; set; }
+        public long RowNumber { get; set; }
 
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
